GeoStalker/requirements.txt
instaloader
requests
Pillow
fpdf
face_recognition
opencv-python
folium
aiohttp             
face_recognition